引入了内存泄漏检测 但是看不太懂内存泄漏 但是不能解决全部内存泄漏问题

+++ app/src/main/java/com/example/ebook_reader/ui/BookShelf/BookShelf.kt
@@ -10,6 +10,7 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.activity.result.PickVisualMediaRequest
 import androidx.activity.result.contract.ActivityResultContracts
+import androidx.appcompat.app.AlertDialog
 import androidx.core.view.isVisible
 import androidx.fragment.app.activityViewModels
 import androidx.lifecycle.lifecycleScope
@@ -26,7 +27,9 @@ import java.io.File
 import java.io.FileOutputStream
 import androidx.core.net.toUri
 import androidx.fragment.app.viewModels
+import com.google.android.material.dialog.MaterialDialogs
 import dagger.hilt.android.AndroidEntryPoint
+import javax.inject.Inject
 
 @AndroidEntryPoint
 class BookShelf : Fragment() {
@@ -36,7 +39,12 @@ class BookShelf : Fragment() {
     //TopBar和BottomBar include布局 和底部抽屉的布局
     private val topICD get() = binding.BookShelfTopBarICD
     private val bottomICD get() = binding.BookshelfBottomBarICD
-    private val bottomSheetDialog = BookShelf_BotSheetDialog()
+    private var _bottomSheetDialog: BookShelf_BotSheetDialog? = BookShelf_BotSheetDialog()
+    private val bottomSheetDialog get() = _bottomSheetDialog!!
+
+    private var bookAdapter: BookAdapter? = null
+    private var alertDialog: AlertDialog?=null
+
 
     //获取Activity共享的ViewModel
     //TODO:通过依赖注入获取ViewModel
@@ -173,7 +181,7 @@ class BookShelf : Fragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        val bookAdapter = initAdapter()
+        initAdapter()
         lifecycleScope.launch {
             //只有在选中一个文件夹时才可以重命名文件夹
             launch {
@@ -191,13 +199,15 @@ class BookShelf : Fragment() {
         }
         updateTopBarVisibility()
         updateBottomBarVisibility()
-        initAllTopICD(bookAdapter)
+        initAllTopICD()
         initAllBottomICD()
     }
-    private fun initAdapter(): BookAdapter{
+
+    private fun initAdapter(){
         //设置书本的ListAdapter
-        //TODO:这里的 bookAdapter应该通过依赖注入来实现
-        val bookAdapter = BookAdapter(viewModel,UIVM)
+        if(bookAdapter==null){
+            bookAdapter = BookAdapter(viewModel,UIVM)
+        }
         //设置layoutManager和adapter
         binding.BookRecyclerView.layoutManager = GridLayoutManager(context, 3)
         binding.BookRecyclerView.adapter = bookAdapter
@@ -206,18 +216,17 @@ class BookShelf : Fragment() {
             launch {
                 viewModel.items.collectLatest { newList ->
                     Log.d("BS 初始化Adapter","initAdapter 更新RecyclerView的显示数据")
-                    bookAdapter.submitList(newList)
+                    bookAdapter!!.submitList(newList)
                 }
             }
             //每次修改EditModule都要更新RecyclerView
             launch {
                 UIVM.isEditModel.collectLatest {
                     Log.d("BS 初始化Adapter","initAdapter 刷新RecyclerView显示数据")
-                    bookAdapter.flashAllViews()
+                    bookAdapter!!.flashAllViews()
                 }
             }
         }
-        return bookAdapter
     }
 
     /**
@@ -274,7 +283,7 @@ class BookShelf : Fragment() {
      * - 新建文件夹按钮
      *
      */
-    private fun initAllTopICD(bookAdapter: BookAdapter){
+    private fun initAllTopICD(){
         //设置在默认页面 Edit模式 进入按钮
         topICD.BookShelfEditBTN.setOnClickListener {
             UIVM.switchEditModel()
@@ -306,12 +315,20 @@ class BookShelf : Fragment() {
         //设置TopBar的全选按钮
         topICD.BookShelfAllSelectBTN.setOnClickListener {
             if (topICD.BookShelfAllSelectBTN.isVisible)
-                bookAdapter.selectedAllSelected()
+                bookAdapter?.selectedAllSelected().also {
+                    if(it==null){
+                        Log.d("BS BookShelfAllSelectBTN","你的Adapter是NULL")
+                    }
+                }
         }
         //设置TopBar的取消全选按钮
         topICD.BookshelfCancelSelectBTN.setOnClickListener {
             if(topICD.BookshelfCancelSelectBTN.isVisible)
-                bookAdapter.cancelAllSelected()
+                bookAdapter?.cancelAllSelected().also{
+                    if(it==null){
+                        Log.d("BS BookshelfCancelSelectBTN","你的Adapter是NULL")
+                    }
+                }
         }
         //设置在文件夹内 重命名文件夹按钮
         topICD.BookShelfRenameFolderInFolderBTN.setOnClickListener {
@@ -336,8 +353,10 @@ class BookShelf : Fragment() {
                 }
 
             }
+
+            alertDialog?.dismiss()
             //弹窗
-            dialogBuilderFactory("新建文件夹",inputNewFolderBoxBinding.root,
+            alertDialog = dialogBuilderFactory("新建文件夹",inputNewFolderBoxBinding.root,
                 {_,_->
                     Log.d("BS 新建文件夹","点击确认")
                     if(inputNewFolderBoxBinding.editTextInput.text?.isEmpty() == true){
@@ -352,26 +371,22 @@ class BookShelf : Fragment() {
                 { _, _ ->
                     Log.d("BS 新建文件夹","点击取消")
                 }
-            )
-                .create()
-                .show()
+            ).also { it.show() }
+
         }
     }
     /**
      * 初始化所有BottomBar的点击事件
      */
     private fun initAllBottomICD(){
-        //设置在编辑模式下 两种页面 移动按钮 呼叫底部抽屉
-        //通过parentFragmentManager管理父 Fragment 或 Activity 中的 Fragment 事务
-        //启动另一个 Fragment（如 DialogFragment）
-        bottomICD.BookShelfMoveBTN.setOnClickListener {
-            bottomSheetDialog.show(parentFragmentManager,bottomSheetDialog.tag )
-
-        }
+        //设置书本移动按钮
+        moveBTNClickListener()
         //设置在编辑模式下 两种页面 删除按钮  添加删除确认弹窗
+
         bottomICD.BookShelfDeleteBTN.setOnClickListener {
+            alertDialog?.dismiss()
             var yourChoice: Boolean? = null
-            dialogBuilderFactory("删除书本或文件夹",
+            alertDialog = dialogBuilderFactory("删除书本或文件夹",
                 "确定删除选中的书本和文件夹吗？(此过程不可逆!)",
                 { _, _ ->
                     deleteSelectedItems()
@@ -382,9 +397,9 @@ class BookShelf : Fragment() {
                     yourChoice = false
                     Log.d("BS DeleteBooksAndFolder","取消删除")
                 }
-            )
-            .create()
-            .show()
+            ).also {
+                it.show()
+            }
             Log.d("BS DeleteBooksAndFolder","$yourChoice")
 
         }
@@ -394,6 +409,30 @@ class BookShelf : Fragment() {
             if(UIVM.isInFolder.value)return@setOnClickListener
 //            Log.d("BS BookShelfRenameFolderBTN","start renameFolder out of Folder")
             showRenameFolderDialog(BookShelfDataViewModel::renameFolder)
+
+        }
+    }
+    /**
+     * 移动书本按钮逻辑
+     * 1. 打开底部抽屉
+     * 2. 显示可以移动的文件夹
+     * 3. 选择一个文件夹
+     *
+     *    ├─ 不选择 选中文件夹为空
+     *
+     *    └─ 选择文件夹 选中文件夹为那个folderId
+     * 4. 点击完成按钮
+     * 5. 把选中的书本修改数据库
+     *
+     *    ├─ 修改后的folderId不存在 不修改数据库
+     *
+     *    └─ 修改后的folderId不存在 修改数据库folderId
+     * 6. 刷新Adapter
+     * 7. 关闭底部抽屉
+     */
+    private fun moveBTNClickListener(){
+        bottomICD.BookShelfMoveBTN.setOnClickListener {
+            bottomSheetDialog.show(parentFragmentManager,bottomSheetDialog.tag )
 
         }
     }
@@ -409,8 +448,8 @@ class BookShelf : Fragment() {
     private fun showRenameFolderDialog(_renameFolder: BookShelfDataViewModel.(String)->Unit ){
         val inputBoxBinding: InputTextboxBinding =
             InputTextboxBinding.inflate(LayoutInflater.from(requireContext()))
-
-        dialogBuilderFactory("文件夹新名称",inputBoxBinding.root,
+        alertDialog?.dismiss()
+        alertDialog = dialogBuilderFactory("文件夹新名称",inputBoxBinding.root,
             {_,_->
                 val newName = inputBoxBinding.editTextInput.text.toString()
                 if (newName.isNotEmpty()){
@@ -423,9 +462,9 @@ class BookShelf : Fragment() {
             { _, _ ->
                 Log.d("BS RenameFolder","取消重命名")
             }
-        )
-        .create()
-        .show()
+        ).also {
+            it.show()
+        }
     }
 
     /**
@@ -438,12 +477,13 @@ class BookShelf : Fragment() {
         negativeButtonClickListener: (DialogInterface, Int) -> Unit,
         positiveButtonText: String = "确定",
         negativeButtonText: String = "取消",
-    ): MaterialAlertDialogBuilder {
+    ): AlertDialog {
         return MaterialAlertDialogBuilder(requireContext())
             .setTitle(title)
             .setView(view)
             .setPositiveButton(positiveButtonText, positiveButtonClickListener)
             .setNegativeButton(negativeButtonText, negativeButtonClickListener)
+            .create()
     }
     /**
      * MaterialAlertDialogBuilder 对话框创建工厂
@@ -455,12 +495,13 @@ class BookShelf : Fragment() {
         negativeButtonClickListener: (DialogInterface, Int) -> Unit,
         positiveButtonText: String = "确定",
         negativeButtonText: String = "取消",
-    ): MaterialAlertDialogBuilder {
+    ): AlertDialog {
         return MaterialAlertDialogBuilder(requireContext())
             .setTitle(title)
             .setMessage(message)
             .setPositiveButton(positiveButtonText, positiveButtonClickListener)
             .setNegativeButton(negativeButtonText, negativeButtonClickListener)
+            .create()
     }
 
     /**
@@ -483,7 +524,14 @@ class BookShelf : Fragment() {
     }
     override fun onDestroy() {
         super.onDestroy()
+
+        binding.BookRecyclerView.adapter=null
         _binding = null
+        bookAdapter = null
+        _bottomSheetDialog?.dismiss()
+        _bottomSheetDialog = null
+        alertDialog?.dismiss()
+        alertDialog=null
         Log.d("BS onDestroy","success")
     }
 }
